requirements.txt
requests
beautifulsoup4
sentence-transformers
qdrant-client
google-generativeai
streamlit
python-dotenv